requirements.txt
requests>=2.28.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
PyPDF2>=3.0.0
python-docx>=0.8.11
markdown>=3.4.0
beautifulsoup4>=4.11.0
jieba>=0.42.1
flask>=2.2.0
flask-cors>=3.0.0
numpy>=1.21.0
